--- src/main/java/roomescape/member/resolver/LoginArgumentResolverConfig.java
@@ -1,21 +1,0 @@
-package roomescape.member.resolver;
-
-import java.util.List;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class LoginArgumentResolverConfig implements WebMvcConfigurer {
-
-    private final LoginMemberArgumentResolver loginMemberArgumentResolver;
-
-    public LoginArgumentResolverConfig(LoginMemberArgumentResolver loginMemberArgumentResolver) {
-        this.loginMemberArgumentResolver = loginMemberArgumentResolver;
-    }
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(loginMemberArgumentResolver);
-    }
-}
